홈 페이징에서 채팅 받으면 채팅이 두 개씩 쌓이던 버그 수정
